third_party/blink/web_tests/svg/custom/clone-node-expected.txt

Tests cloneNode for SVGDocument.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS svgDoc.cloneNode(false).__proto__ is XMLDocument.prototype
PASS className(svgDoc.cloneNode(false)) is "XMLDocument"
PASS svgDoc.cloneNode(true).compatMode is "CSS1Compat"
PASS successfullyParsed is true

TEST COMPLETE
